Related to #3909 - moved extensions to Testing namespace, some others to Internals
