+++ src/Transports/MassTransit.Azure.ServiceBus.Core/Testing/AzureServiceBusTestHarness.cs
@@ -5,6 +5,7 @@
     using System.Threading.Tasks;
     using Azure;
     using Azure.Messaging.ServiceBus.Administration;
+    using Internals;
 
 
     public class AzureServiceBusTestHarness :
