+++ tests/MassTransit.Tests/KillSwitch_Specs.cs
@@ -3,6 +3,7 @@ namespace MassTransit.Tests
     using System;
     using System.Linq;
     using System.Threading.Tasks;
+    using MassTransit.Testing;
     using NUnit.Framework;
     using TestFramework;
